% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch12p10.R
\name{ch12p10}
\alias{ch12p10}
\title{Chapter 12.3 Problem#10: Pulse Rates}
\format{A data frame with 40 rows and 3 variables}
\source{
Triola, M. 2014. Elementary Statistics, 12th ed. Boston, MA: Pearson.
}
\description{
Pulse rates (bpm) by age category (under or over 30) and sex.  
\itemize{
\item pulse.rate
\item Sex
\item Age (categorical as over or under 30)
}
}
