% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch12p6.R
\name{ch12p6}
\alias{ch12p6}
\title{Chapter 12.3 Problem#6: Chest Deceleration (in g)}
\format{A data frame with 18 rows and 3 variables}
\source{
Triola, M. 2014. Elementary Statistics, 12th ed. Boston, MA: Pearson.
}
\description{
Chest deceleration by car type and size.  
\itemize{
\item Chest.Deceleration (in grams)
\item Type (Foreign or Domestic)
\item Size (Small, Medium, Large)
}
}
