% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds02_foot_height.R
\name{ds02_foot_height}
\alias{ds02_foot_height}
\title{Data Set 2: Foot and Height Measurements}
\format{A data frame with 40 rows and 6 variables}
\source{
Triola, M. 2018. Elementary Statistics, 13th ed. Boston, MA: Pearson.
}
\description{
A data set containing data on 40 foot and height measurements.  Data is from "Estimation of Stature from
Foot Length: Applications in Forensic Science, " by Brenda Rohren, M.A., MFS, LIMHP, LADS, MAC.
Brenda Rohren was a graduate student at Nebraska Wesleyan University when she conducted the
research and wrote the report.
\itemize{
\item SEX is gender of subject
\item AGE is in years
\item FOOT_LENGTH foot length (cm)
\item SHOE_SIZE is reported shoe size
\item HEIGHT is is height (cm) of the subject
}
}
