% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds03_body_temperatures.R
\name{ds03_body_temperatures}
\alias{ds03_body_temperatures}
\title{Data Set 3: Body Temperatures (in degrees Fahrenheit) of Healthy Adults}
\format{A data frame with 107 rows and 6 variables.}
\source{
Triola, M. 2018. Elementary Statistics, 13th ed. Boston, MA: Pearson.
}
\description{
A data set containing 107 adult body temperatures taken on two consecutive days at 8am and 12am.  Data provided by
Dr. Steven Wasserman, Dr. Philips Mackowiak, and Dr. Myron
Levine of the University of Maryland.
\itemize{
\item SEX is the gender of the subject
\item SMOKE indicates if subject smokes (Y) or does not smoke (N)
\item DAY_1_8AM is the 8am body temperature (in degrees Fahrenheit) of an
adult on the first day of the research study.
\item DAY_1_12AM is the 12am body temperature (in degrees Fahrenheit) of an
adult on the first day of the research study.
\item DAY_2_8AM is the 8am body temperature (in degrees Fahrenheit) of an
adult on the second day of the research study.
\item DAY_2_12AM is the 12am body temperature (in degrees Fahrenheit) of an
adult on the second day of the research study.
}
}
