% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds04_births.R
\name{ds04_births}
\alias{ds04_births}
\title{Data Set 4: Births}
\format{A data frame with 400 rows and 8 variables}
\source{
Triola, M. 2018. Elementary Statistics, 13th ed. Boston, MA: Pearson.
}
\description{
Data are from 400 births.  For GENDER 1 = male and 0 = female.
LENGTH_OF_STAY is in days, BIRTH_WEIGHT is in grams, and TOTAL_CHARGES
are in dollars.
\itemize{
\item FACILITY
\item INSURANCE
\item GENDER (1 = Male)
\item LENGTH_OF_STAY (days)
\item ADMITTED
\item DISCHARGED
\item BIRTH_WEIGHT (grams)
\item TOTAL_CHARGES (dollars)
}
}
