% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds05_family_heights.R
\name{ds05_family_heights}
\alias{ds05_family_heights}
\title{Data Set 5: Family Heights}
\format{A data frame with 134 rows and 4 variables}
\source{
Triola, M. 2018. Elementary Statistics, 13th ed. Boston, MA: Pearson.
}
\description{
Height data are from 134 families.  Heights are in inches.  Only families
with at least one child of each gender are included, and only heights
of the first son and first daughter are included.  The data are from
a journalof Francis Galton (1822-1911), who developed the concepts of
standard deviation, regression line, and correlation between two
variables
\itemize{
\item FATHER
\item MOTHER
\item FIRST_SON
\item FIRST_DAUGHTER
}
}
