% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds06_freshman_15.R
\name{ds06_freshman_15}
\alias{ds06_freshman_15}
\title{Data Set 6: Freshman 15 Data}
\format{A data frame with 67 rows and 5 variables}
\source{
Triola, M. 2018. Elementary Statistics, 13th ed. Boston, MA: Pearson.
}
\description{
Measurements were made in September of freshman year and then
later in April of freshman year.  Results are published in "Changes
in Body Weight and Fat Mass of Men and Women in the First Year
of College: A Study of the 'Freshman 15' " by Hoffman, Policastro,
Quick, and Lee, Journal of American College Health, Vol. 55, No. 1.
\itemize{
\item SEX is the gender of the subject
\item WT_SEPT is the September weight (kg)
\item WT_APRIL is the April weight (kg)
\item BMI_SEPT denotes measured body mass index in September
\item BMI_APRIL denotes measured body mass index in April
}
}
