% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds10.R
\name{ds10}
\alias{ds10}
\title{Data Set 10: Cigarette Tar, Nicotine, and Carbon Monoxide}
\format{A data frame with 25 rows and 9 variables}
\source{
Triola, M. 2014. Elementary Statistics, 12th ed. Boston, MA: Pearson.
}
\description{
All measurements are in milligrams per cigarette.  CO denotes carbon monoxide.
The king-size cigarettes are nonfiltered, nonmenthol, and nonlight.  The 
menthol cigarettes are 100 mm long, filtered, and nonlight.  The cigarettes
in the third group are 100 mm long, filtered, nonmenthol, and nonlight.  Data
are from the Federal Trade Commission.
\itemize{
\item KgTar King size tar
\item KgNic King size nicotine
\item KgCO  King size carbon monoxide
\item MnTar Menthol tar
\item MnNic Menthol nicotine
\item MnCO Menthol carbon monoxide
\item FLTar Filtered tar
\item FLNic Filtered nicotine
\item FLCO Filtered carbon monoxide  
}
}

