% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds14_Oscar_Winner_Age.R
\name{ds14_Oscar_Winner_Age}
\alias{ds14_Oscar_Winner_Age}
\title{Data Set 14: Ages of Oscar Winners}
\format{A data frame with 87 rows and 2 variables}
\source{
Triola, M. 2018. Elementary Statistics, 13th ed. Boston, MA: Pearson.
}
\description{
Ages (years) of actresses and actors at the times that they won Oscars for the
categories of Best Actress and Best Actor.  The ages are listed in chronological
order by row, so that corresponding locations in the two tables are from the
same year.  (Notes: In 1968 there was a tie in the Best Actress category, and
the mean of the two ages is used; in 1932 there was a tie in the best Actor
category, and the mean of the two ages is used.  These data are suggested by
the article "Ages of Oscar-winning Best Actors and Actresses" by Richard Brown
and Gretchen Davis, \strong{Mathematics Teacher} magazine.  In that article,
the year of birth of the award winner was subtracted from the year the awards
ceremony, but the ages listed here are calculated from teh birth date of the
winner and the date of the awards ceremony.) The data are complete as of this
writing.
\itemize{
\item Actresses Age of Best Actress winners
\item Actors Age of Best Actor winners.
}
}
