% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds15.R
\name{ds15}
\alias{ds15}
\title{Data Set 15: Flight Data}
\format{A data frame with 48 rows and 6 variables}
\source{
Triola, M. 2014. Elementary Statistics, 12th ed. Boston, MA: Pearson.
}
\description{
All flights are American Airlines flights from New York (JFK) to Los Angeles
(LAX), and all flights occurred in January of a recent year.  Data are from the
Bureau of Transportation.
\itemize{
\item Plane.ID is the identification number on the tail of the aircraft
\item Dep.Delay is the departure delay time (minutes), and negative numbers corresponds to flights that departed early.
\item Taxi.Out is the time (minutes) that the flight used to taxi from the terminal to the runway for departure.
\item Taxi.In is the time (minutes) that the flight used to taxi to the terminal after landing.
\item Arr.Delay is the arrival delay time (minutes), and negative numbers correspond to flights that arrived early (before the scheduled arrival time).
}
}

