% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds16.R
\name{ds16}
\alias{ds16}
\title{Data Set 16: Earthquake Measurements}
\format{A data frame with 50 rows and 2 variables}
\source{
Triola, M. 2014. Elementary Statistics, 12th ed. Boston, MA: Pearson.
}
\description{
Fifty matched pairs of megnitude/depth measurements randomly selected from 10,594 
earthquakes recorded in one year from a location in southern California.  The
magnitude and depth both describe the source of the earthquake.  The data from the 
Southern California Erathquake Data Center.
\itemize{
\item MAG is magnitude measured on the Richter scale.
\item DEPTH is depth in km.
}
}
