% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds16_Nobel_Laureates_and_Chocolate.R
\name{ds16_Nobel_Laureates_and_Chocolate}
\alias{ds16_Nobel_Laureates_and_Chocolate}
\title{Data Set 16: Nobel Laureates and Chocolate}
\format{A data frame with 23 rows and 5 variables}
\source{
Triola, M. 2018. Elementary Statistics, 13th ed. Boston, MA: Pearson.
}
\description{
Data are from 23 countries.  The data are from "The
Real Secret to Genius? Reading Between the Lines, " by
McClintock, Stangle, and Cetinkaya-Rundel, \strong{Chance},
Vol. 27, No. 1; and "Chocolate Consumption, Cognitive Function,
and Nobel Laureates," by Franz Messerli, \strong{New England Journal of Medicine},
Vol. 367, No. 16.
\itemize{
\item COUNTRY
\item CHOCOLATE includes chocolate consumption (kg per capita)
\item NOBEL includes the number of Nobel Laureates (per 10 million people)
\item POPULATION includes population (in millions)
\item INTERNET includes the number of Internet users per 100 people
}
}
