% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds17.R
\name{ds17}
\alias{ds17}
\title{Data Set 17: Word Counts by Males and Females}
\format{A data frame with 56 rows and 12 variables}
\source{
Triola, M. 2014. Elementary Statistics, 12th ed. Boston, MA: Pearson.
}
\description{
The columns are counts of the number of words spoken in a day by male (M) and
female (F) subjects in six different sample groups.  Column M1 denotes the
word counts for males in Sample 1, F1 is the count for females in Sample 1, and
so on.
\itemize{
\item Sample 1: Recruited couples ranging in age from 18 to 29
\item Sample 2: Students recruited in introductory psychology classes, aged 17 to 23
\item Sample 3: Students recruited in introductory psychology classes in Mexico, aged 17 to 25
\item Sample 4: Students recruited in introductory psychology classes, aged 17 to 22
\item Sample 5: Students recruited in introductory psychology classes, aged 18 to 26
\item Sample 6: Students recruited in introductory psychology classes, aged 17 to 23
}
}
