% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds18_Speed_Dating.R
\name{ds18_Speed_Dating}
\alias{ds18_Speed_Dating}
\title{Data Set 18: Speed Dating}
\format{A data frame with 199 rows and 10 variables}
\source{
Triola, M. 2018. Elementary Statistics, 13th ed. Boston, MA: Pearson.
}
\description{
Data are from 199 dates.  Based on replication data from \strong{Data Analysis Using Regression and Multilevel/Hierarchical Models},
by Andrew Gelman and Jennifer Hill, Cambridge University Press.  Data for males uses the corresponding decsriptors are defined for females.
\itemize{
\item DEC_BY_FEM is decision (1 = yes) of female to date again
\item AGE_FEM is age of female
\item LIKE_BY_FEM is "like" rating by female of male (scale of 1-10)
\item ATTRACT_BY_FEM is "attractive" rating by female of male (scale 1-10)
\item ATTRIB_BY_FEM is sum of ratinsg of fove attributes (sincerity, intelligence, fun, ambitious, shared interests) by female of male
}
}
