% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds19.R
\name{ds19}
\alias{ds19}
\title{Data Set 19: Cola Weights and Volumes}
\format{A data frame with 36 rows and 8 variables}
\source{
Triola, M. 2014. Elementary Statistics, 12th ed. Boston, MA: Pearson.
}
\description{
A data set containing data on cola weights and volumes.  Weights are in pounds and volumes are in ounces.
\itemize{
\item CRGWT weight regular coke
\item CRGVL volume regular coke
\item CDTWT weight diet coke
\item CDTVL volume diet coke
\item PRGWT weight regular pepsi
\item PRGVL volume regular pepsi
\item PDTWT weight diet pepsi
\item PDTVL volume diet pepsi
}
}

