% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds19_Car_Crash_Tests.R
\name{ds19_Car_Crash_Tests}
\alias{ds19_Car_Crash_Tests}
\title{Data Set 19: Car Crash Tests}
\format{A data frame with 21 rows and 8 variables}
\source{
Triola, M. 2018. Elementary Statistics, 13th ed. Boston, MA: Pearson.
}
\description{
The same cars are used in Data Set 20.  The data are measurements from cars
crashed into a fixed barrier at 35 mi/h with a crash test dummy in the driver's
seat.  Data are from the National Highway Traffic Safety Administration.
\itemize{
\item HIC is a measurement of a standard "head injury criterion".
\item CHEST is chest deceleration (in g, where g is a force of gravity)
\item FEMUR_L is the measured load on the left femur (in lb)
\item FEMUR_R is the meausred load on the right femur (in lb)
\item TTI is a measurement of the side thoracic trauma index
\item PELVIS is pelvis deceleration (in g, where g is a force of gravity)
}
}
