% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds2.R
\name{ds2}
\alias{ds2}
\title{Data Set 2: Foot and Height Measurements}
\format{A data frame with 40 rows and 6 variables}
\source{
Triola, M. 2014. Elementary Statistics, 12th ed. Boston, MA: Pearson.
}
\description{
A data set containing data on 40 foot and height measurements.  Data is from "Estimation of Stature from
Foot Length: Applications in Forensic Science, " by Brenda rohren, M.A., MFS, LIMHP, LADS, MAC.
Brenda Rohren was a graduate student at Nebraska Wesleyan University when she conducted the
research and wrote the report.
\itemize{
\item Sex is gender of subject
\item Age is in years
\item Foot.Length foot length (cm)
\item Shoe.Size is reported shoe size
\item Height is is height (cm) of the subject
}
}

