% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds20.R
\name{ds20}
\alias{ds20}
\title{Data Set 20: M&M Plain Candy Weights (grams)}
\format{A data frame with 27 rows and 6 variables}
\source{
Triola, M. 2014. Elementary Statistics, 12th ed. Boston, MA: Pearson.
}
\description{
M&M candy weights in grams by color.  
\itemize{
\item Red
\item Orange
\item Yellow
\item Brown
\item Blue
\item Green
}
}

