% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds20_Car_Measurements.R
\name{ds20_Car_Measurements}
\alias{ds20_Car_Measurements}
\title{Data Set 20: Car Measurements}
\format{A data frame with 21 rows and 10 variables}
\source{
Triola, M. 2018. Elementary Statistics, 13th ed. Boston, MA: Pearson.
}
\description{
The same cars are used in Data Set 19.  The data are measurements from cars
that have automatic transmissions and were manufactured in the same recent year.
\itemize{
\item WEIGHT is weight (lb)
\item LENGTH is length (inches)
\item BRAKING is braking distance (feet) from 60 mi/h
\item CYLINDERS is the number of cylinders
\item DISPLACEMENT is the engine displacement (liters)
\item CITY is the fuel comsumption (mi/gal) for city driving conditions
\item HIGHWAY is the fuel comsumption (mi/gal) for highway driving conditions
\item GHG is a measure of greenhouse gas emissions (in tons/year, expressed as CO2 equivalents)
}
}
