% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds21.R
\name{ds21}
\alias{ds21}
\title{Data Set 21: Coin Weights (grams)}
\format{A data frame with 40 rows and 8 variables}
\source{
Triola, M. 2014. Elementary Statistics, 12th ed. Boston, MA: Pearson.
}
\description{
The "pre-1983 pennies" were made after the Indian and wheat pennies, and they are 97\% copper and 3\% zinc.  The "post-1983 pennies" are 3\% copper and 97\% zinc.  The "pre-1964 silver quarters" are 90\% silver and 10\% copper. The 'post-1964 quarters" are made with a copper-nickel alloy.
\itemize{
\item Indian.Pennies
\item Wheat.Pennies
\item Pre.1983.Pennies
\item Post.1983.Pennies
\item Canadian.Pennies
\item Pre.1964.Quarters
\item Post.1964.Quarters
\item Dollar.Coins
}
}
