% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds21_earthquakes.R
\name{ds21_earthquakes}
\alias{ds21_earthquakes}
\title{Data Set 21: Earthquakes}
\format{A data frame with 600 rows and 2 variables}
\source{
Triola, M. 2018. Elementary Statistics, 13th ed. Boston, MA: Pearson.
}
\description{
Data are from 600 matches pairs magnitude/depth measurements randomly selected
from  10,594 earthquakes recorded in one year from a location in southern
California.  Only earthquakes with a magnitude of at least 1.00 are used.
}
\details{
\itemize{
\item MAGNITUDE is magnitude measured on the Richter scale
\item DEPTH is depth in km.  The magnitude and depth both describe the source of the earthquake.  The data are from the Southern California Earthquake Data Center.
}
}
