% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds22.R
\name{ds22}
\alias{ds22}
\title{Data Set 22: Axial Loads of Aluminum Cans}
\format{A data frame with 175 rows and 2 variables}
\source{
Triola, M. 2014. Elementary Statistics, 12th ed. Boston, MA: Pearson.
}
\description{
Axial load are measured in pounds.  Axial loads are applied when the tops are pressed into place.
\itemize{
\item CN109
\item CN111
}
}

