% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds22_Tornadoes.R
\name{ds22_Tornadoes}
\alias{ds22_Tornadoes}
\title{Data Set 22: Tornadoes}
\format{A data frame with 500 rows and 6 variables}
\source{
Triola, M. 2018. Elementary Statistics, 13th ed. Boston, MA: Pearson.
}
\description{
Data are from 500 tornadoes arranged chronologically.
Data are from the National Weather Service.
}
\details{
\itemize{
\item MONTH is the month of the tornado (1 = January)
\item F_SCALE is the Fujita scale rating of tornado intensity.
\item FATALITIES is number of deaths caused by the tornado
\item LENGTH is the distance the tornado traveled in miles
\item WIDTH is the tornado width in yards.
}
}
