% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds23.R
\name{ds23}
\alias{ds23}
\title{Data Set 23: Weights of Discarded Garbage for One Week}
\format{A data frame with 62 rows and 10 variables}
\source{
Triola, M. 2014. Elementary Statistics, 12th ed. Boston, MA: Pearson.
}
\description{
Weights are in pounds.  Data provided by Masakuza Tani, the Garbage Project, University of Arizona.
\itemize{
\item HHSize is the household size
\item Metal
\item Paper
\item Plastic
\item Glass
\item Food
\item Yard
\item Text
\item Other
\item Total
}
}
