% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds23_Old_Faithful.R
\name{ds23_Old_Faithful}
\alias{ds23_Old_Faithful}
\title{Data Set 23: Old Faithful}
\format{A data frame with 250 rows and 5 variables}
\source{
Triola, M. 2018. Elementary Statistics, 13th ed. Boston, MA: Pearson.
}
\description{
Data are from 250 eruptions of the Old Faithful geyser in Yellowstone National Park.
Based on data from the Geyser Observation and Study Association.
\itemize{
\item INT_BEFORE is the time interval (min) before the eruption
\item DURATION is the time (sec) of the eruption
\item INT_AFTER is the time interval (min) after the eruption
\item HEIGHT (ft) is the height of the eruption
\item PRED_ERROR is the error (min) of he predicted time of eruption
}
}
