% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds25_Fast_Food.R
\name{ds25_Fast_Food}
\alias{ds25_Fast_Food}
\title{Data Set 25: Fast Food}
\format{A data frame with 50 rows and 8 variables}
\source{
Triola, M. 2018. Elementary Statistics, 13th ed. Boston, MA: Pearson.
}
\description{
Data are from 400 observations of drive-thru service times (sec)
at different fast-food restaurants.  Times begin when a vehicle
stops at the order window and end when the vehicle leaves the
pickup window.  Lunch times were measured between 11:00 AM and 2:00 PM,
and dinner times were measured between 4:00 PM and 7:00 PM.  Data collected
by the author.
}
\details{
\itemize{
\item MCDONALDS
\item BURGER KING
\item WENDYS
\item DUNKIN DONUTS
}
}
