% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds26_Cola_Weights_and_Volumes.R
\name{ds26_Cola_Weights_and_Volumes}
\alias{ds26_Cola_Weights_and_Volumes}
\title{Data Set 26: Cola Weights and Volumes}
\format{A data frame with 36 rows and 8 variables}
\source{
Triola, M. 2018. Elementary Statistics, 13th ed. Boston, MA: Pearson.
}
\description{
A data set containing data on cola weights and volumes.  Weights are in pounds and volumes are in ounces.
\itemize{
\item COKE_REG_WT weight regular coke
\item COKE_REG_VOL volume regular coke
\item COKE_DIET_WT weight diet coke
\item COKE_DIET_VOL volume diet coke
\item PEPSI_REG_WT weight regular pepsi
\item PEPSI_REG_VOL volume regular pepsi
\item PEPSI_DIET_WT weight diet pepsi
\item PEPSI_DIET_VOL volume diet pepsi
}
}
