% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds27_M_M_Weights.R
\name{ds27_M_M_Weights}
\alias{ds27_M_M_Weights}
\title{Data Set 27: M&M Plain Candy Weights (grams)}
\format{A data frame with 27 rows and 6 variables}
\source{
Triola, M. 2018. Elementary Statistics, 13th ed. Boston, MA: Pearson.
}
\description{
M&M candy weights in grams by color.
\itemize{
\item Red
\item Orange
\item Yellow
\item Brown
\item Blue
\item Green
}
}
