% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds3.R
\name{ds3}
\alias{ds3}
\title{Data Set 3: Body Temperatures (in degrees Fahrenheit) of Healthy Adults}
\format{A data frame with 107 rows and 1 variable}
\source{
Triola, M. 2014. Elementary Statistics, 12th ed. Boston, MA: Pearson.
}
\description{
A data set containing 107 adult body temperatures taken at 12am.  Data provided by
Dr. Steven Wasserman, Dr. Philips Mackowiak, and Dr. Myron
Levine of the University of Maryland.
\itemize{
\item Temperature is the 12am body temperature (in degrees Fahrenheit) of an 
adult on the second day of a research study.
}
}
