% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds5.R
\name{ds5}
\alias{ds5}
\title{Data Set 5: IQ and Lead Exposure}
\format{A data frame with 121 rows and 8 variables}
\source{
Triola, M. 2014. Elementary Statistics, 12th ed. Boston, MA: Pearson.
}
\description{
Data are measured from children in two consecutive years, and the 
children were living close to a lead smelter.
Data are from "Neuropsychological Dysfunction in Children with 
Chronic Low-Level Lead Absorption, " by P.J. Landrigan, R.H.
Whitworth, R.W. Baloh, N.W. Staehling, W.F. Barthel, and B.F. Rosenblum,
Lancet, Vol. 1, Issue 7909.
\itemize{
\item LEAD is blood lead level group [1=low lead level (blood lead
levels < 40 micograms/100 mL in both years; 2 = medium lead level (blood lead levels >= 40 micrograms/100 mL
in exactly one of the two years); 3 = high lead levels (blood lead levels >= 40 micrograms/100 mL
in both years)].
\item AGE is age in years
\item SEX is sex of subject (1=male,2=female)
\item YEAR1 is blood lead level in first year
\item YEAR2 is blood lead level in second year
\item IQV is measured verbal IQ score
\item IQP is measured performance IQ score
\item IQF is measured full IQ score
}
}

