% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds6.R
\name{ds6}
\alias{ds6}
\title{Data Set 6: IQ and Brain Size}
\format{A data frame with 20 rows and 9 variables}
\source{
Triola, M. 2014. Elementary Statistics, 12th ed. Boston, MA: Pearson.
}
\description{
The data are obtained from monozygotic (identical) twins.  Data provided by M.J.Tramo,
W.C. loftus, T.A. Stukel, J.B. Weaver, M.S. Gazziniga.  See "Brain Size, Head Size, and IQ
in Monozygotic Twins, " Neurology, Vol. 50.
\itemize{
\item PAIR identifies the set of twins
\item SEX is the gender of the subject (1 = male, 2 = female)
\item ORDER is the birth order
\item IQ is full-scale IQ
\item VOL is total brain volumne (cm^3)
\item AREA is total brain surface area (cm^2)
\item CCSA is corpus callsum (fissure connecting left and right crebral hemispheres)
surface area (cm^2)
\item CIRC is head circumference (cm)
\item WT is body weight (kg)
}
}
