% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds8.R
\name{ds8}
\alias{ds8}
\title{Data Set 8: Alcohol and Tobacco Use in Animated Children's Movies}
\format{A data frame with 50 rows and 5 variables}
\source{
Triola, M. 2014. Elementary Statistics, 12th ed. Boston, MA: Pearson.
}
\description{
The data are based on "Tobacco and Alcohol Use in G-Rated Children's
Animated films," by Goldstein, Sobel, and Newman, Journal of the
American Medical Association, Vol. 281, No. 12.
\itemize{
\item Movie name
\item Company 
\item Length of the movie in minutes
\item Tobacco.Use in the movie in seconds
\item Alcohol.Use in the movie in seconds
}
}

